requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyPDF2==3.0.1
google-generativeai==0.3.1
requests==2.31.0
gunicorn==21.2.0
werkzeug==3.0.1
